+++ artichoke-backend/src/extn/core/regexp/mod.rs
@@ -9,7 +9,7 @@ use std::hash::{Hash, Hasher};
 use std::mem;
 use std::rc::Rc;
 
-use crate::convert::{Convert, RustBackedValue};
+use crate::convert::{Convert, Int, RustBackedValue};
 use crate::def::{rust_data_free, ClassLike, Define};
 use crate::eval::Eval;
 use crate::extn::core::error::{RubyException, RuntimeError, SyntaxError, TypeError};
@@ -146,23 +146,23 @@ impl RustBackedValue for Regexp {
     fn new_obj_args(&self, interp: &Artichoke) -> Vec<sys::mrb_value> {
         vec![
             Value::convert(interp, self.literal_pattern.as_str()).inner(),
-            Value::convert(interp, self.literal_options.flags().bits()).inner(),
+            Value::convert(interp, self.literal_options.flags().bits() as i32).inner(),
             Value::convert(interp, self.encoding.flags()).inner(),
         ]
     }
 }
 
 impl Regexp {
-    pub const IGNORECASE: i64 = 1;
-    pub const EXTENDED: i64 = 2;
-    pub const MULTILINE: i64 = 4;
+    pub const IGNORECASE: Int = 1;
+    pub const EXTENDED: Int = 2;
+    pub const MULTILINE: Int = 4;
 
-    pub const ALL_REGEXP_OPTS: i64 = Self::IGNORECASE | Self::EXTENDED | Self::MULTILINE;
+    pub const ALL_REGEXP_OPTS: Int = Self::IGNORECASE | Self::EXTENDED | Self::MULTILINE;
 
-    pub const FIXEDENCODING: i64 = 16;
-    pub const NOENCODING: i64 = 32;
+    pub const FIXEDENCODING: Int = 16;
+    pub const NOENCODING: Int = 32;
 
-    pub const ALL_ENCODING_OPTS: i64 = Self::FIXEDENCODING | Self::NOENCODING;
+    pub const ALL_ENCODING_OPTS: Int = Self::FIXEDENCODING | Self::NOENCODING;
 
     pub fn new(
         literal_pattern: String,
